Add gtest_main_manifest target to avoid duplicates of main.cpp files for test applications - use it in frontends tests
