+++ src/core/tests/frontend/onnx/CMakeLists.txt
@@ -8,7 +8,8 @@ file(GLOB SRC ${CMAKE_CURRENT_SOURCE_DIR}/*.cpp)
 
 add_executable(${TARGET_NAME} ${SRC})
 
-target_link_libraries(${TARGET_NAME} PRIVATE frontend_shared_test_classes openvino_onnx_frontend frontend_common)
+target_link_libraries(${TARGET_NAME}
+        PRIVATE frontend_shared_test_classes openvino_onnx_frontend frontend_common gtest_main_manifest)
 
 add_clang_format_target(${TARGET_NAME}_clang FOR_TARGETS ${TARGET_NAME})
