+++ src/frontends/tensorflow/tests/CMakeLists.txt
@@ -8,7 +8,8 @@ file(GLOB SRC ${CMAKE_CURRENT_SOURCE_DIR}/*.cpp)
 
 add_executable(${TARGET_NAME} ${SRC})
 
-target_link_libraries(${TARGET_NAME} PRIVATE frontend_shared_test_classes openvino_tensorflow_frontend openvino::runtime)
+target_link_libraries(${TARGET_NAME} PRIVATE
+        gtest_main_manifest frontend_shared_test_classes openvino_tensorflow_frontend openvino::runtime)
 
 add_clang_format_target(${TARGET_NAME}_clang FOR_TARGETS ${TARGET_NAME})
