--- src/frontends/tensorflow/tests/main.cpp
@@ -1,9 +1,0 @@
-// Copyright (C) 2018-2022 Intel Corporation
-// SPDX-License-Identifier: Apache-2.0
-//
-
-#include "utils.hpp"
-
-int main(int argc, char** argv) {
-    return FrontEndTestUtils::run_tests(argc, argv);
-}
