+++ src/frontends/tests/frontend/shared/src/utils.cpp
@@ -35,3 +35,15 @@ std::string FrontEndTestUtils::find_ov_path() {
 #    error "Unsupported OS"
 #endif
 }
+
+std::string FrontEndTestUtils::get_disabled_tests(const std::string& manifest_path) {
+    std::string result = "-";
+    std::ifstream manifest_stream(manifest_path);
+
+    for (std::string line; std::getline(manifest_stream, line);) {
+        if (line.size() && (line[0] != '#')) {
+            result += ":" + line;
+        }
+    }
+    return result;
+}
